Refactor LED tests and monitoring scripts
