+++ led-testing/gnss_sma_monitor.py
@@ -1,334 +1,290 @@
 #!/usr/bin/env python3
 """
-GNSS и SMA мониторинг с LED индикацией для TimeCard
+Мониторинг GNSS и SMA статусов с LED индикацией
 Автор: AI Assistant
-
-Цветовая схема:
-- Зеленый: GNSS SYNC или SMA сигналы достоверны
-- Красный: GNSS LOST или SMA сигналы недостоверны  
-- Сиреневый: Режим holdover (автономное хранение)
-- Желтый: Промежуточные состояния
 """
 
-import os
-import time
-import json
 import subprocess
-from datetime import datetime
-from typing import Dict, Any, Optional
+import json
+import time
+import sys
+import os
 
-class GNSSSMAMonitor:
+class GNSSSMAMonitorFixed:
     def __init__(self, bus=1, addr=0x37):
         self.bus = bus
         self.addr = addr
-        self.timecard_sysfs = "/sys/class/timecard/ocp0"
+        self.quantum_pci_timecard_sysfs = "/sys/class/timecard/ocp0"
         
-        # LED индексы для разных статусов
+        # ИСПРАВЛЕННАЯ схема нумерации LED
+        # LED 1,3,5 используют регистры 0x01,0x03,0x05
+        # LED 2,4,6 используют регистры 0x07,0x09,0x0B
         self.led_map = {
-            'gnss_sync': 0,      # Power LED - зеленый/красный для GNSS
-            'gnss_holdover': 1,   # Sync LED - сиреневый для holdover
-            'sma3_status': 2,     # GNSS LED - статус SMA3
-            'sma4_status': 3,     # Alarm LED - статус SMA4
-            'clock_source': 4,    # Status1 LED - источник часов
-            'system_status': 5    # Status2 LED - общий статус системы
+            'gnss_sync': 0,      # Power LED - регистр 0x01
+            'gnss_holdover': 1,   # Sync LED - регистр 0x07
+            'sma3_status': 2,     # GNSS LED - регистр 0x03
+            'sma4_status': 3,     # Alarm LED - регистр 0x09
+            'clock_source': 4,    # Status1 LED - регистр 0x05
+            'system_status': 5    # Status2 LED - регистр 0x0B
         }
         
-        # Цветовые коды для IS32FL3207
+        # Маппинг LED индексов на правильные регистры
+        self.led_registers = {
+            0: 0x01,  # LED 1 - Power
+            1: 0x07,  # LED 2 - Sync (исправлено!)
+            2: 0x03,  # LED 3 - GNSS
+            3: 0x09,  # LED 4 - Alarm (исправлено!)
+            4: 0x05,  # LED 5 - Status1
+            5: 0x0B   # LED 6 - Status2 (исправлено!)
+        }
+        
         self.colors = {
             'off': 0x00,
-            'green': 0xFF,        # Полная яркость для зеленого
-            'red': 0xFF,          # Полная яркость для красного  
-            'purple': 0x80,       # Средняя яркость для сиреневого
-            'yellow': 0xC0,       # Высокая яркость для желтого
-            'dim_green': 0x40,    # Тусклый зеленый
-            'dim_red': 0x40,      # Тусклый красный
-            'blink': 0x60         # Мигающий режим
+            'green': 0xFF,
+            'red': 0xFF,
+            'purple': 0x80,
+            'yellow': 0xC0
         }
         
-        # Статусы для мониторинга
-        self.status_cache = {}
-        
-    def check_i2c_device(self) -> bool:
-        """Проверка наличия IS32FL3207"""
+        # Инициализация контроллера
+        self.init_led_controller()
+    
+    def check_i2c_device(self):
+        """Проверка наличия I2C устройства"""
         try:
             result = subprocess.run(['sudo', 'i2cdetect', '-y', str(self.bus)], 
                                   capture_output=True, text=True, check=True)
-            return "37" in result.stdout
+            # Ищем адрес в формате "37" (как отображается в i2cdetect)
+            return f"{self.addr:02x}" in result.stdout
         except subprocess.CalledProcessError:
             return False
     
-    def init_led_controller(self) -> bool:
+    def init_led_controller(self):
         """Инициализация LED контроллера"""
         try:
             # Включение контроллера
-            subprocess.run(['sudo', 'i2cset', '-y', str(self.bus), str(self.addr), '0x00', '0x01'], 
-                         check=True)
+            subprocess.run(['sudo', 'i2cset', '-y', str(self.bus), f'0x{self.addr:02x}', '0x00', '0x01'], check=True)
             
             # Установка глобального тока
-            subprocess.run(['sudo', 'i2cset', '-y', str(self.bus), str(self.addr), '0x6E', '0xFF'], 
-                         check=True)
+            subprocess.run(['sudo', 'i2cset', '-y', str(self.bus), f'0x{self.addr:02x}', '0x6E', '0xFF'], check=True)
             
             # Настройка scaling регистров
             for i in range(18):
-                subprocess.run(['sudo', 'i2cset', '-y', str(self.bus), str(self.addr), 
-                              f'0x{0x4A + i:02x}', '0xFF'], check=True)
+                reg = 0x4A + i
+                subprocess.run(['sudo', 'i2cset', '-y', str(self.bus), f'0x{self.addr:02x}', f'0x{reg:02x}', '0xFF'], check=True)
             
-            # Обновление
-            subprocess.run(['sudo', 'i2cset', '-y', str(self.bus), str(self.addr), '0x49', '0x00'], 
-                         check=True)
+            # Обновление контроллера
+            subprocess.run(['sudo', 'i2cset', '-y', str(self.bus), f'0x{self.addr:02x}', '0x49', '0x00'], check=True)
             
+            print("✅ LED контроллер инициализирован")
             return True
-        except subprocess.CalledProcessError:
+        except subprocess.CalledProcessError as e:
+            print(f"❌ Ошибка инициализации LED контроллера: {e}")
             return False
     
-    def set_led_brightness(self, led_index: int, brightness: int) -> bool:
-        """Установка яркости LED"""
+    def set_led_brightness(self, led_index, brightness):
+        """Установка яркости LED с правильным регистром"""
+        if led_index not in self.led_registers:
+            print(f"❌ Неизвестный LED индекс: {led_index}")
+            return False
+        
         try:
-            # PWM регистр (0x01-0x23 для LED 1-18)
-            pwm_reg = 0x01 + led_index
-            subprocess.run(['sudo', 'i2cset', '-y', str(self.bus), str(self.addr), 
-                          f'0x{pwm_reg:02x}', f'0x{brightness:02x}'], check=True)
-            
-            # Обновление
-            subprocess.run(['sudo', 'i2cset', '-y', str(self.bus), str(self.addr), '0x49', '0x00'], 
-                         check=True)
+            reg = self.led_registers[led_index]
+            subprocess.run(['sudo', 'i2cset', '-y', str(self.bus), f'0x{self.addr:02x}', f'0x{reg:02x}', f'0x{brightness:02x}'], check=True)
+            subprocess.run(['sudo', 'i2cset', '-y', str(self.bus), f'0x{self.addr:02x}', '0x49', '0x00'], check=True)
             return True
-        except subprocess.CalledProcessError:
+        except subprocess.CalledProcessError as e:
+            print(f"❌ Ошибка установки LED {led_index}: {e}")
             return False
     
-    def read_sysfs_value(self, path: str) -> Optional[str]:
+    def read_sysfs_value(self, attribute):
         """Чтение значения из sysfs"""
         try:
-            if os.path.exists(path):
-                with open(path, 'r') as f:
-                    return f.read().strip()
-        except Exception:
-            pass
-        return None
+                            with open(f"{self.quantum_pci_timecard_sysfs}/{attribute}", 'r') as f:
+                return f.read().strip()
+        except (FileNotFoundError, PermissionError):
+            return None
     
-    def get_gnss_status(self) -> Dict[str, Any]:
+    def get_gnss_status(self):
         """Получение статуса GNSS"""
+        gnss_sync = self.read_sysfs_value('gnss_sync')
+        clock_source = self.read_sysfs_value('clock_source')
+        
         status = {
-            'sync': self.read_sysfs_value(f"{self.timecard_sysfs}/gnss_sync"),
-            'clock_source': self.read_sysfs_value(f"{self.timecard_sysfs}/clock_source"),
-            'available_sources': self.read_sysfs_value(f"{self.timecard_sysfs}/available_clock_sources"),
-            'drift': self.read_sysfs_value(f"{self.timecard_sysfs}/clock_status_drift"),
-            'offset': self.read_sysfs_value(f"{self.timecard_sysfs}/clock_status_offset")
+            'sync': gnss_sync == 'SYNC' if gnss_sync else False,
+            'holdover': clock_source == 'holdover' if clock_source else False,
+            'source': clock_source if clock_source else 'unknown'
         }
         
-        # Определение режима работы
-        if status['sync'] == 'SYNC':
-            status['mode'] = 'sync'
-            status['color'] = 'green'
-        elif status['sync'] == 'LOST':
-            status['mode'] = 'lost'
-            status['color'] = 'red'
-        elif status['clock_source'] in ['MAC', 'IRIG-B', 'external']:
-            status['mode'] = 'holdover'
-            status['color'] = 'purple'
-        else:
-            status['mode'] = 'unknown'
-            status['color'] = 'yellow'
-        
         return status
     
-    def get_sma_status(self) -> Dict[str, Any]:
+    def get_sma_status(self):
         """Получение статуса SMA выходов"""
-        status = {}
+        sma_status = {}
         
         for i in range(1, 5):
-            sma_path = f"{self.timecard_sysfs}/sma{i}"
-            sma_value = self.read_sysfs_value(sma_path)
-            
+            sma_value = self.read_sysfs_value(f'sma{i}')
             if sma_value:
-                # Определение достоверности сигнала
-                if 'OUT:' in sma_value:
-                    # Выходной сигнал - проверяем источник
-                    if 'PHC' in sma_value or '10Mhz' in sma_value:
-                        status[f'sma{i}'] = {
-                            'type': 'output',
-                            'signal': sma_value,
-                            'reliable': True,
-                            'color': 'green'
-                        }
-                    else:
-                        status[f'sma{i}'] = {
-                            'type': 'output', 
-                            'signal': sma_value,
-                            'reliable': False,
-                            'color': 'red'
-                        }
-                else:
-                    # Входной сигнал
-                    status[f'sma{i}'] = {
-                        'type': 'input',
-                        'signal': sma_value,
-                        'reliable': True,
-                        'color': 'green'
-                    }
+                sma_status[f'sma{i}'] = sma_value
             else:
-                status[f'sma{i}'] = {
-                    'type': 'unknown',
-                    'signal': 'N/A',
-                    'reliable': False,
-                    'color': 'red'
-                }
+                sma_status[f'sma{i}'] = 'unknown'
         
-        return status
+        return sma_status
     
-    def update_led_status(self, gnss_status: Dict, sma_status: Dict):
-        """Обновление LED индикации"""
-        
-        # GNSS Sync LED (Power LED)
-        if gnss_status['mode'] == 'sync':
-            self.set_led_brightness(self.led_map['gnss_sync'], self.colors['green'])
-        elif gnss_status['mode'] == 'lost':
-            self.set_led_brightness(self.led_map['gnss_sync'], self.colors['red'])
-        else:
-            self.set_led_brightness(self.led_map['gnss_sync'], self.colors['yellow'])
-        
-        # Holdover LED (Sync LED)
-        if gnss_status['mode'] == 'holdover':
-            self.set_led_brightness(self.led_map['gnss_holdover'], self.colors['purple'])
-        else:
-            self.set_led_brightness(self.led_map['gnss_holdover'], self.colors['off'])
-        
-        # SMA3 Status LED (GNSS LED)
-        if 'sma3' in sma_status:
-            if sma_status['sma3']['reliable']:
-                self.set_led_brightness(self.led_map['sma3_status'], self.colors['green'])
-            else:
-                self.set_led_brightness(self.led_map['sma3_status'], self.colors['red'])
-        
-        # SMA4 Status LED (Alarm LED)
-        if 'sma4' in sma_status:
-            if sma_status['sma4']['reliable']:
-                self.set_led_brightness(self.led_map['sma4_status'], self.colors['green'])
-            else:
-                self.set_led_brightness(self.led_map['sma4_status'], self.colors['red'])
-        
-        # Clock Source LED (Status1 LED)
-        if gnss_status['clock_source'] == 'GNSS':
-            self.set_led_brightness(self.led_map['clock_source'], self.colors['green'])
-        elif gnss_status['clock_source'] in ['MAC', 'IRIG-B']:
-            self.set_led_brightness(self.led_map['clock_source'], self.colors['purple'])
-        else:
-            self.set_led_brightness(self.led_map['clock_source'], self.colors['yellow'])
-        
-        # System Status LED (Status2 LED)
-        overall_status = 'green'
-        if gnss_status['mode'] == 'lost':
-            overall_status = 'red'
-        elif gnss_status['mode'] == 'holdover':
-            overall_status = 'purple'
-        elif not all(sma.get('reliable', False) for sma in sma_status.values()):
-            overall_status = 'yellow'
-        
-        self.set_led_brightness(self.led_map['system_status'], self.colors[overall_status])
-    
-    def turn_off_all_leds(self):
-        """Выключение всех LED"""
-        for i in range(18):
-            self.set_led_brightness(i, self.colors['off'])
-    
-    def export_metrics(self, filename: str = None) -> Dict[str, Any]:
-        """Экспорт метрик в JSON"""
+    def update_led_status(self):
+        """Обновление статуса LED на основе GNSS и SMA"""
         gnss_status = self.get_gnss_status()
         sma_status = self.get_sma_status()
         
-        metrics = {
-            "timestamp": datetime.now().isoformat(),
-            "device": "TimeCard GNSS/SMA Monitor",
-            "gnss": gnss_status,
-            "sma": sma_status,
-            "led_status": {
-                "gnss_sync": gnss_status.get('color', 'unknown'),
-                "holdover": 'purple' if gnss_status['mode'] == 'holdover' else 'off',
-                "sma3": sma_status.get('sma3', {}).get('color', 'unknown'),
-                "sma4": sma_status.get('sma4', {}).get('color', 'unknown'),
-                "clock_source": gnss_status.get('color', 'unknown'),
-                "system": 'green' if all(sma.get('reliable', False) for sma in sma_status.values()) else 'red'
+        # LED 0: GNSS Sync Status (Power LED)
+        if gnss_status['sync']:
+            self.set_led_brightness(0, self.colors['green'])  # Зеленый - SYNC
+        else:
+            self.set_led_brightness(0, self.colors['red'])    # Красный - LOST
+        
+        # LED 1: GNSS Holdover Status (Sync LED)
+        if gnss_status['holdover']:
+            self.set_led_brightness(1, self.colors['purple']) # Сиреневый - holdover
+        else:
+            self.set_led_brightness(1, self.colors['off'])    # Выключен
+        
+        # LED 2: SMA3 Status (GNSS LED)
+        # Проверяем что SMA3 работает как выход И GNSS синхронизирован
+        sma3_working = sma_status.get('sma3', '').startswith('OUT:')
+        if gnss_status['sync'] and sma3_working:
+            self.set_led_brightness(2, self.colors['green'])  # Зеленый - надежный
+        else:
+            self.set_led_brightness(2, self.colors['red'])    # Красный - ненадежный
+        
+        # LED 3: SMA4 Status (Alarm LED)
+        # Проверяем что SMA4 работает как выход И GNSS синхронизирован
+        sma4_working = sma_status.get('sma4', '').startswith('OUT:')
+        if gnss_status['sync'] and sma4_working:
+            self.set_led_brightness(3, self.colors['green'])  # Зеленый - надежный
+        else:
+            self.set_led_brightness(3, self.colors['red'])    # Красный - ненадежный
+        
+        # LED 4: Clock Source Status (Status1 LED)
+        if gnss_status['source'] == 'gnss':
+            self.set_led_brightness(4, self.colors['green'])  # Зеленый - GNSS
+        elif gnss_status['source'] == 'holdover':
+            self.set_led_brightness(4, self.colors['purple']) # Сиреневый - holdover
+        else:
+            self.set_led_brightness(4, self.colors['yellow']) # Желтый - другой источник
+        
+        # LED 5: System Status (Status2 LED)
+        if gnss_status['sync'] and not gnss_status['holdover']:
+            self.set_led_brightness(5, self.colors['green'])  # Зеленый - нормальная работа
+        else:
+            self.set_led_brightness(5, self.colors['yellow']) # Желтый - предупреждение
+        
+        return {
+            'gnss': gnss_status,
+            'sma': sma_status,
+            'leds': {
+                'power': 'green' if gnss_status['sync'] else 'red',
+                'sync': 'purple' if gnss_status['holdover'] else 'off',
+                'gnss': 'green' if (gnss_status['sync'] and sma_status.get('sma3', '').startswith('OUT:')) else 'red',
+                'alarm': 'green' if (gnss_status['sync'] and sma_status.get('sma4', '').startswith('OUT:')) else 'red',
+                'status1': 'green' if gnss_status['source'] == 'gnss' else ('purple' if gnss_status['source'] == 'holdover' else 'yellow'),
+                'status2': 'green' if (gnss_status['sync'] and not gnss_status['holdover']) else 'yellow'
             }
         }
-        
-        if filename:
-            with open(filename, 'w') as f:
-                json.dump(metrics, f, indent=2)
-        
-        return metrics
     
-    def monitor_loop(self, interval: int = 5):
+    def turn_off_all_leds(self):
+        """Выключение всех LED"""
+        print("🔚 Выключение всех LED...")
+        for led_index in range(6):
+            self.set_led_brightness(led_index, 0)
+        print("✅ Все LED выключены")
+    
+    def export_metrics(self, status):
+        """Экспорт метрик в JSON файл"""
+        metrics = {
+            'timestamp': time.time(),
+            'gnss_sync': status['gnss']['sync'],
+            'gnss_holdover': status['gnss']['holdover'],
+            'clock_source': status['gnss']['source'],
+            'sma_status': status['sma'],
+            'led_status': status['leds']
+        }
+        
+        try:
+            with open('gnss_sma_metrics_fixed.json', 'w') as f:
+                json.dump(metrics, f, indent=2)
+        except Exception as e:
+            print(f"❌ Ошибка экспорта метрик: {e}")
+    
+    def print_status(self, status):
+        """Вывод статуса в консоль"""
+        print("\n" + "="*60)
+        print("📊 СТАТУС GNSS И SMA МОНИТОРИНГА")
+        print("="*60)
+        
+        # GNSS статус
+        gnss = status['gnss']
+        print(f"🛰️  GNSS Sync: {'✅ SYNC' if gnss['sync'] else '❌ LOST'}")
+        print(f"🔄 Holdover: {'✅ АКТИВЕН' if gnss['holdover'] else '❌ НЕ АКТИВЕН'}")
+        print(f"📡 Источник часов: {gnss['source']}")
+        
+        # SMA статус
+        print("\n🔌 SMA Выходы:")
+        for sma, value in status['sma'].items():
+            # Определяем статус надежности для SMA3 и SMA4
+            if sma in ['sma3', 'sma4']:
+                is_output = value.startswith('OUT:')
+                reliability = "✅ НАДЕЖНЫЙ" if (status['gnss']['sync'] and is_output) else "❌ НЕНАДЕЖНЫЙ"
+                print(f"  {sma.upper()}: {value} ({reliability})")
+            else:
+                print(f"  {sma.upper()}: {value}")
+        
+        # LED статус
+        print("\n💡 LED Индикация:")
+        led_names = ['Power', 'Sync', 'GNSS', 'Alarm', 'Status1', 'Status2']
+        for i, (name, color) in enumerate(zip(led_names, status['leds'].values())):
+            print(f"  {name}: {color}")
+        
+        print("="*60)
+    
+    def monitor_loop(self, interval=5):
         """Основной цикл мониторинга"""
-        print("=== GNSS/SMA Monitor для TimeCard ===")
-        print(f"Интервал обновления: {interval} секунд")
-        print("Нажмите Ctrl+C для остановки")
-        print()
+        print("🚀 Запуск мониторинга GNSS и SMA статусов...")
+        print("💡 Используется ИСПРАВЛЕННАЯ схема нумерации LED")
+        print("🎯 LED 2,4,6 используют регистры 0x07,0x09,0x0B")
         
         try:
             while True:
-                # Получение статусов
-                gnss_status = self.get_gnss_status()
-                sma_status = self.get_sma_status()
-                
-                # Обновление LED
-                self.update_led_status(gnss_status, sma_status)
-                
-                # Вывод статуса
-                self.print_status(gnss_status, sma_status)
-                
-                # Экспорт метрик
-                self.export_metrics("gnss_sma_metrics.json")
-                
+                status = self.update_led_status()
+                self.export_metrics(status)
+                self.print_status(status)
                 time.sleep(interval)
                 
         except KeyboardInterrupt:
             print("\n🛑 Остановка мониторинга...")
             self.turn_off_all_leds()
-            print("✅ Все LED выключены")
-            print("🎉 Мониторинг завершен")
-    
-    def print_status(self, gnss_status: Dict, sma_status: Dict):
-        """Вывод текущего статуса"""
-        timestamp = datetime.now().strftime("%H:%M:%S")
-        
-        print(f"[{timestamp}] === Статус GNSS/SMA ===")
-        
-        # GNSS статус
-        print(f"GNSS Sync: {gnss_status['sync']} ({gnss_status['color']})")
-        print(f"Clock Source: {gnss_status['clock_source']}")
-        print(f"Mode: {gnss_status['mode']}")
-        
-        # SMA статус
-        print("SMA Status:")
-        for sma_name, sma_data in sma_status.items():
-            status_icon = "🟢" if sma_data['reliable'] else "🔴"
-            print(f"  {sma_name.upper()}: {status_icon} {sma_data['signal']}")
-        
-        print()
+            print("✅ Мониторинг остановлен")
 
 def main():
-    monitor = GNSSSMAMonitor()
+    """Главная функция"""
+    monitor = GNSSSMAMonitorFixed()
     
-    # Проверка TimeCard
-    if not os.path.exists(monitor.timecard_sysfs):
-        print("❌ TimeCard не найден")
-        return
-    
-    print(f"✅ TimeCard найден: {monitor.timecard_sysfs}")
-    
-    # Проверка I2C устройства
     if not monitor.check_i2c_device():
-        print("❌ IS32FL3207 не найден")
-        return
+        print("❌ IS32FL3207 не найден на I2C шине")
+        sys.exit(1)
     
     print("✅ IS32FL3207 найден")
     
-    # Инициализация LED контроллера
-    if not monitor.init_led_controller():
-        print("❌ Ошибка инициализации LED контроллера")
-        return
+    # Демонстрация работы
+    print("\n🎯 Демонстрация исправленной схемы LED:")
     
-    print("✅ LED контроллер инициализирован")
+    # Тест каждого LED
+    for i in range(6):
+        print(f"Тест LED {i} (регистр 0x{monitor.led_registers[i]:02x})...")
+        monitor.set_led_brightness(i, monitor.colors['green'])
+        time.sleep(1)
+        monitor.set_led_brightness(i, 0)
+    
+    print("\n✅ Демонстрация завершена")
     
     # Запуск мониторинга
     monitor.monitor_loop()
