--- ДРАЙВЕРА/.Module.symvers.cmd
@@ -1,1 +1,0 @@
-savedcmd_Module.symvers :=  /usr/src/linux-headers-6.14.0-24-generic/scripts/mod/modpost -M -m -b -x -a      -o Module.symvers -T modules.order -i /usr/src/linux-headers-6.14.0-24-generic/Module.symvers -e
