--- ДРАЙВЕРА/.modules.order.cmd
@@ -1,1 +1,0 @@
-savedcmd_modules.order := {   echo ptp_ocp.o; :; } > modules.order
