--- ДРАЙВЕРА/.ptp_ocp.ko.cmd
@@ -1,1 +1,0 @@
-savedcmd_ptp_ocp.ko := ld -r -m elf_x86_64 -z noexecstack --no-warn-rwx-segments --build-id=sha1  -T /usr/src/linux-headers-6.14.0-24-generic/scripts/module.lds -o ptp_ocp.ko ptp_ocp.o ptp_ocp.mod.o .module-common.o
