--- ДРАЙВЕРА/modules.order
@@ -1,1 +1,0 @@
-ptp_ocp.o
